Update puma
Security alert:
GHSA-33vf-4xgg-9r58
moderate severity
Vulnerable versions: >= 4.0.0, < 4.3.3
Patched version: 4.3.3
Impact
If an application using Puma allows untrusted input in an early-hints
header, an attacker can use a carriage return character to end the
header and inject malicious content, such as additional headers or an
entirely new response body. This vulnerability is known as HTTP Response
Splitting

While not an attack in itself, response splitting is a vector for
several other attacks, such as cross-site scripting (XSS).

This is related to CVE-2020-5247, which fixed this vulnerability but
only for regular responses.